1. Added VAT 2. Added QR Code 3. Multiuser login bug fix

+++ src/main/java/org/ideoholic/curium/model/feescollection/dto/FeesCategoryResponseDto.java
@@ -19,11 +19,11 @@ public class FeesCategoryResponseDto {
     private String dayOne;
     private String dateFrom;
     private String dateTo;
-    private Map<String, Long> feeCategoryCollectionMapReport;
-    private long feesByCash;
-    private long feesByBank;
-    private long feesByCashOtherFees;
-    private long feesByBankOtherFees;
+    private Map<String, Double> feeCategoryCollectionMapReport;
+    private double feesByCash;
+    private double feesByBank;
+    private double feesByCashOtherFees;
+    private double feesByBankOtherFees;
     private boolean success;
 
 }
